Add files via upload
These are the header files and source files for a program that converts between rectangular and spherical coordinates. The main.cpp file is a test file provided for the lab that uses the spherical.cpp and spherical.hpp written by me to test the conversion functions are within the tolerance. 
Input: Nothing 
Return: Passed or Failed for 7 different tests. 

Adam Claxton